+++ NEWS.md
@@ -8,6 +8,7 @@
 * `removeEmptyValLabels()` for removing unused missing tags and value labels (#4)
 * `extractData()` and `extractData2()` now correctly apply value labels, even when value and value label conflicts exist
 * `write_spss2()` now also handles variables containing only NA values when `format` is also NA (#72)
+* `changeMissings()` now correctly changes missing labels for values of variables with partially non-existent `value`s and/or `valLabel`s (#73)
 
 # eatGADS 1.1.0